create folder for db and sort getUsers()
